requirements.txt
abstra==2.25.4
pandas==2.2.2
requests==2.31.0
python-docx==1.1.2
Unidecode==1.3.8
docxtpl==0.18.0
docusign-esign==4.0.0rc1
python-dotenv==1.0.1
loguru==0.7.2
pypandoc==1.13